Use same category selector everywhere
