+++ src/pages/registration/resource_type_tab/SelectRegistrationTypeField.tsx
@@ -1,25 +1,16 @@
 import CloseIcon from '@mui/icons-material/Close';
 import FilterVintageIcon from '@mui/icons-material/FilterVintage';
-import SearchIcon from '@mui/icons-material/Search';
-import { Box, Chip, FormHelperText, FormLabel, IconButton, Paper, TextField, Typography } from '@mui/material';
+import { Box, Chip, FormHelperText, FormLabel, IconButton, Paper, Typography } from '@mui/material';
 import { ErrorMessage, useFormikContext } from 'formik';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
+import { CategorySelector } from '../../../components/CategorySelector';
 import { ConfirmDialog } from '../../../components/ConfirmDialog';
 import { RootState } from '../../../redux/store';
 import {
-  ArtisticType,
-  BookType,
-  ChapterType,
   DegreeType,
-  ExhibitionContentType,
-  JournalType,
-  MediaType,
-  OtherRegistrationType,
-  PresentationType,
   PublicationType,
-  ReportType,
   ResearchDataType,
   ResourceFieldNames,
   contextTypeBaseFieldName,
@@ -57,11 +48,9 @@ import { PublicationChannelType, PublicationInstanceType, Registration } from '.
 import { dataTestId } from '../../../utils/dataTestIds';
 import {
   getMainRegistrationType,
-  isDegreeWithProtectedFiles,
   isPeriodicalMediaContribution,
   nviApplicableTypes,
 } from '../../../utils/registration-helpers';
-import { RegistrationTypeElement, RegistrationTypesRow } from './components/RegistrationTypesRow';
 
 export const SelectRegistrationTypeField = () => {
   const { t } = useTranslation();
@@ -69,15 +58,11 @@ export const SelectRegistrationTypeField = () => {
   const { values, setFieldValue, validateForm } = useFormikContext<Registration>();
   const currentInstanceType = values.entityDescription?.reference?.publicationInstance?.type ?? '';
 
-  const [searchValue, setSearchValue] = useState('');
   const [openSelectType, setOpenSelectType] = useState(!currentInstanceType);
   const [confirmNewType, setConfirmNewType] = useState<PublicationInstanceType | ''>('');
   const [showDatasetConditions, setShowDatasetConditions] = useState(false);
 
-  const closeSelectType = () => {
-    setOpenSelectType(false);
-    setSearchValue('');
-  };
+  const closeSelectType = () => setOpenSelectType(false);
 
   const updateRegistrationData = (newInstanceType: PublicationInstanceType) => {
     if (newInstanceType !== currentInstanceType) {
@@ -227,13 +212,6 @@ export const SelectRegistrationTypeField = () => {
     }
   };
 
-  const filterRegistrationTypes = (registrationTypes: RegistrationTypeElement[]) => {
-    const lowerCaseSearchValue = searchValue.toLowerCase();
-    return registrationTypes.filter((registrationType) =>
-      registrationType.text.toLowerCase().includes(lowerCaseSearchValue)
-    );
-  };
-
   return openSelectType || !currentInstanceType ? (
     <>
       <Paper sx={{ p: '1rem' }} elevation={10}>
@@ -248,170 +226,17 @@ export const SelectRegistrationTypeField = () => {
             </IconButton>
           )}
         </Box>
-        <Box
-          sx={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            flexWrap: 'wrap',
-            gap: '0.5rem',
-            my: '1rem',
-          }}>
-          <TextField
-            data-testid={dataTestId.registrationWizard.resourceType.resourceTypeSearchField}
-            sx={{ maxWidth: '17rem' }}
-            type="search"
-            variant="filled"
-            label={t('common.search')}
-            InputProps={{
-              endAdornment: <SearchIcon />,
-            }}
-            onChange={(event) => setSearchValue(event.target.value)}
-          />
-          <Box
-            sx={{
-              display: 'flex',
-              gap: '0.5rem',
-              alignItems: 'center',
-            }}>
-            <FilterVintageIcon
-              color="primary"
-              titleAccess={t('registration.resource_type.nvi.can_give_publication_points')}
-              fontSize="small"
-            />
-            <Typography>{t('registration.resource_type.nvi.can_give_publication_points')}</Typography>
-          </Box>
-        </Box>
 
-        <Box
-          sx={{
-            display: 'grid',
-            gridTemplateColumns: { xs: '1fr', md: 'auto 1fr' },
-            gap: '1rem',
-            alignItems: 'center',
-          }}>
-          <RegistrationTypesRow
-            mainType={PublicationType.PublicationInJournal}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(JournalType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Book}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(BookType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Report}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(ReportType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Degree}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(DegreeType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-                disabled: isDegreeWithProtectedFiles(registrationType) && !user?.isThesisCurator,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Anthology}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(ChapterType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Presentation}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(PresentationType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.Artistic}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(ArtisticType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.MediaContribution}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(MediaType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.ResearchData}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(ResearchDataType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.ExhibitionContent}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(ExhibitionContentType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-          <RegistrationTypesRow
-            mainType={PublicationType.GeographicalContent}
-            registrationTypes={filterRegistrationTypes(
-              Object.values(OtherRegistrationType).map((registrationType) => ({
-                value: registrationType,
-                text: t(`registration.publication_types.${registrationType}`),
-                selected: registrationType === currentInstanceType,
-              }))
-            )}
-            onChangeType={onChangeType}
-          />
-        </Box>
+        <CategorySelector
+          selectedCategories={currentInstanceType ? [currentInstanceType] : []}
+          onCategoryClick={onChangeType}
+          disabledCategories={
+            !user?.isThesisCurator
+              ? [DegreeType.Bachelor, DegreeType.Master, DegreeType.Phd, DegreeType.Other]
+              : undefined
+          }
+        />
+
         {!currentInstanceType && (
           <FormHelperText error sx={{ mt: '1rem' }}>
             <ErrorMessage name={ResourceFieldNames.RegistrationType} />
