+++ src/pages/search/advanced_search/AdvancedSearchPage.tsx
@@ -4,12 +4,12 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
 import { FetchResultsParams, ResultParam, SortOrder, fetchResults } from '../../../api/searchApi';
+import { CategoryChip } from '../../../components/CategorySelector';
 import { SearchForm } from '../../../components/SearchForm';
 import { SortSelector } from '../../../components/SortSelector';
 import { RegistrationFieldName } from '../../../types/publicationFieldNames';
 import { PublicationInstanceType } from '../../../types/registration.types';
 import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
-import { CategoryChip } from '../../registration/resource_type_tab/components/RegistrationTypesRow';
 import { ExportResultsButton } from '../ExportResultsButton';
 import { PublicationDateIntervalFilter } from '../PublicationDateIntervalFilter';
 import { RegistrationSearch } from '../registration_search/RegistrationSearch';
